+++ R/shadow-package.R
@@ -24,4 +24,5 @@
 #' @import sp
 #' @import raster
 #' @importFrom methods as
+
 NULL